requirements.txt
azure-storage==0.37.0
nltk==3.8.1
numpy==2.3.3
openai==1.109.1
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openslide-python==1.3.1
pandas==2.3.2
Pillow==11.3.0
pyautogen==0.2.26
requests==2.32.5
rouge-score==0.1.2
scikit-learn==1.3.2
scikit-survival==0.22.2
scikit-image==0.22.0
timm==1.0.20
torch==2.0.0+cu117
torchvision==0.15.0+cu117
tqdm==4.66.5
